moved ClickEvent to FavoriteTradesOverlay.java and made all the things local instead of static
